+++ bin/replicante/src/main.rs
@@ -17,7 +17,7 @@ fn main() {
     match result {
         Err(error) => {
             let message = format_fail(&error);
-            println!("{}", message);
+            eprintln!("{}", message);
             exit(1);
         }
         Ok(clean) if !clean => exit(1),