+++ models/core/src/api/mod.rs
@@ -1,6 +1,9 @@
 use serde_derive::Deserialize;
 use serde_derive::Serialize;
 
+pub mod apply;
+pub mod validate;
+
 /// Replicante version information.
 #[derive(Clone, Eq, PartialEq, Hash, Debug, Serialize, Deserialize)]
 pub struct Version {